feat(testament crud): rename services
