+++ backend/src/app.module.ts
@@ -10,7 +10,7 @@ import { ConfigModule, ConfigService } from '@nestjs/config'
 import { ScheduleModule } from '@nestjs/schedule'
 import { AuthModule } from './auth/auth.module'
 import { DbModule } from './db/db.module'
-import { LastWillModule } from './last-wills/lastwill.module'
+import { LastWillModule } from './lastwill/lastwill.module'
 import { MailModule } from './mail/mail.module'
 import { PaymentsModule } from './payments/payments.module'
 import { ProfileModule } from './profile/profile.module'
