+++ backend/src/db/db.module.ts
@@ -1,13 +1,13 @@
 import { TypegooseModule } from '@m8a/nestjs-typegoose'
 import { Module } from '@nestjs/common'
 import { MODELS } from './entities'
-import { LastWillDBService } from './services/lastwill.db.service'
-import { MailEventService } from './services/mail-event.service'
-import { UserService } from './services/user.service'
+import { LastWillDBService } from './services/lastwill.service'
+import { MailEventDBService } from './services/mail-event.service'
+import { UserDBService } from './services/user.service'
 
 @Module({
   imports: [TypegooseModule.forFeature([...MODELS])],
-  providers: [UserService, MailEventService, LastWillDBService],
-  exports: [UserService, MailEventService, LastWillDBService],
+  providers: [UserDBService, MailEventDBService, LastWillDBService],
+  exports: [UserDBService, MailEventDBService, LastWillDBService],
 })
 export class DbModule {}
