+++ backend/src/db/services/lastwill.service.ts
@@ -2,8 +2,8 @@ import { InjectModel } from '@m8a/nestjs-typegoose'
 import { Injectable, Logger, NotFoundException } from '@nestjs/common'
 import { ReturnModelType } from '@typegoose/typegoose'
 import { ObjectId } from 'mongoose'
-import { CreateLastWillDto } from '../../last-wills/dto/create-lastwill.dto'
-import { UpdateLastWillDto } from '../../last-wills/dto/update-lastwill.dto'
+import { CreateLastWillDto } from '../../lastwill/dto/create-lastwill.dto'
+import { UpdateLastWillDto } from '../../lastwill/dto/update-lastwill.dto'
 import { LastWill } from '../entities/lastwill.entity'
 
 @Injectable()
