+++ backend/test/lastwills.e2e-spec.ts
@@ -8,7 +8,7 @@ import * as request from 'supertest'
 import { DbModule } from '../src/db/db.module'
 import { LastWill, LastWillMetadata } from '../src/db/entities/lastwill.entity'
 import { User } from '../src/db/entities/users.entity'
-import { LastWillModule } from '../src/last-wills/lastwill.module'
+import { LastWillModule } from '../src/lastwill/lastwill.module'
 import { paymentPlans } from '../src/payments/interfaces/payments'
 import { SharedModule } from '../src/shared/shared.module'
 import { MockConfigService } from './helpers/config-service.helper'
